+++ src/index.mobile.js
@@ -1,5 +1,0 @@
-import './assets/styles/globalStyling.scss';
-
-export * from './lib/atoms/index.mobile';
-export * from './lib/molecules/index.mobile';
-export * from './lib/organisms/index.mobile';