+++ src/lib/atoms/Avatar/index.tsx
@@ -1,0 +1,1 @@
+export { IAvatarProps, default as default } from './Avatar';