+++ src/lib/molecules/Charts/PieChart/index.js
@@ -4,8 +4,9 @@ import HighchartsReact from 'highcharts-react-official';
 import PropTypes from 'prop-types';
 import classnames from 'classnames';
 import { useDeviceType } from 'hooks';
+import { Empty, BusyLoader } from 'components';
 import ChartLegend from '../Legend';
-import { Empty, BusyLoader } from '../../../atoms';
+
 
 import './index.scss';