+++ src/lib/organisms/Form/Formables/FormableDropdown.js
@@ -1,4 +1,4 @@
-import FormableHOC from '../FormableHOC';
+import DropdownField from '../FormableHOC';
 import ValidatableDropdown from '../../../molecules/ValidatableElements/Elements/ValidatableDropdown';
 
-export default FormableHOC(ValidatableDropdown);
+export default DropdownField(ValidatableDropdown);