+++ src/lib/organisms/Form/Formables/FormableRadio.js
@@ -1,4 +1,4 @@
-import FormableHOC from '../FormableHOC';
+import Radio from '../FormableHOC';
 import ValidatableRadio from '../../../molecules/ValidatableElements/Elements/ValidatableRadio';
 
-export default FormableHOC(ValidatableRadio);
+export default Radio(ValidatableRadio);