+++ src/lib/organisms/Form/Formables/FormableSwitcher.js
@@ -1,4 +1,4 @@
-import FormableHOC from '../FormableHOC';
+import SwitcherElement from '../FormableHOC';
 import ValidatableSwitcher from '../../../molecules/ValidatableElements/Elements/ValidatableSwitcher';
 
-export default FormableHOC(ValidatableSwitcher);
+export default SwitcherElement(ValidatableSwitcher);