+++ src/lib/organisms/Table/Row/col.js
@@ -2,10 +2,10 @@ import React, { useRef, useCallback, useEffect, useState } from 'react';
 import classnames from 'classnames';
 import isEqual from 'react-fast-compare';
 
-import { guid, stopEvent, copyToClipboard } from 'utils';
-import Icon from '../../../atoms/Icon';
-import SkeletonLoader from '../../../atoms/SkeletonLoader';
-import Tooltip from '../../../molecules/Tooltip';
+import { guid, stopEvent, copyToClipboard, callAfterDelay } from 'utils';
+import Icon from 'src/lib/atoms/Icon';
+import SkeletonLoader from 'src/lib/atoms/SkeletonLoader';
+import Tooltip from 'src/lib/molecules/Tooltip';
 
 function Col({
     id,
@@ -25,6 +25,7 @@ function Col({
     copyableValue,
     stickyColumns,
     copyTooltipText,
+    copiedTooltipText,
     initialColWidth,
     disabledColumnPin,
     defaultCustomWidth,
@@ -53,6 +54,7 @@ function Col({
             return guidRef.current;
         }
     });
+    const [isCopied, setIsCopied] = useState(false);
 
     useEffect(() => {
         mounted.current = true;
@@ -70,8 +72,12 @@ function Col({
         (event) => {
             stopEvent(event);
             copyToClipboard(copyableValue || value);
+            !isCopied && setIsCopied(true);
+            callAfterDelay(() => {
+                setIsCopied(false);
+            }, 2000);
         },
-        [copyableValue, value]
+        [copyableValue, value, isCopied]
     );
 
     return (
@@ -97,12 +103,12 @@ function Col({
             >
                 <SkeletonLoader height={20} isBusy={guidRef.current && promiseValue === guidRef.current}>
                     {copyable && value && (
-                        <Tooltip title={copyTooltipText}>
+                        <Tooltip title={isCopied ? copiedTooltipText : copyTooltipText}>
                             <Icon
                                 tabIndex="1"
                                 className="cursor-pointer copy-icon"
-                                type="bc-icon-copy-mirror"
-                                onClick={handleCopy}
+                                type={isCopied ? 'bc-icon-checkbox-checked' : 'bc-icon-copy-mirror'}
+                                onClick={!isCopied && handleCopy}
                             />
                         </Tooltip>
                     )}
@@ -116,7 +122,8 @@ function Col({
 }
 
 Col.defaultProps = {
-    copyTooltipText: 'Copy'
+    copyTooltipText: 'Copy',
+    copiedTooltipText: 'Copied!'
 };
 
 export default React.memo(Col, isEqual);